requirements.txt
Flask~=2.0.1
Werkzeug~=2.0.1
python-magic==0.4.24
Flask-SQLAlchemy~=2.5.1
SQLAlchemy-Utils~=0.37.8
psycopg2-binary~=2.9.1